Initial Commit
Add the current, working code. This solution does require a single external library, memmap2. This is to make use the (g)libc mmap syscall without having to generate the ASM/libc bindings myself.

The "next_newline" function shares much of the logic with the Java implementation from the world record, however there were some adaptations to make it work for rust. This is mainly in cases where we would read past the end of the MMap buffer. The Java version would zero-out any bytes after if it was a partial read, however rust will not do that.