Reduce runtime by 25ms
Reduce the time to run the program by 25ms by having the program spawn a child process. The child will do all the processing the pipe the output to the parent, and close STDOUT. This signals to the parent to read all the data and exit. Effectively, this removes the need for the munmap syscall to be counted as part of the program runtime