Add Support for 10K City Challenge
tivrfoa pointed out that the main branch only covers the smaller data set, as opposed to the larger one. While the main branch works, it is less performant because of the required re-allocations on the hashmap.

The README has been updated with timing info for the 10k city challenge as opposed to the smaller test set