Fix Off by One Error
When the reading a line, and we could not get exactly 8 bytes we return the end of the file. However, the file ends in a newline. Because of this, when it was passed to the 'parse_line' function it would incorrectly use the newline as an integer (0x0A). As a result, the JSON output did not match that of other people